Feature/remove convex log store (#46)
* refactor: remove redundant Convex log storage

* chore: remove internal development plan

* feat: add CI checks and refresh Convex tooling

Remove keepalive workflow and obsolete task logging backend while adding
formatting, linting, typechecking, build scripts, and expanded Convex agent
skills.

* fix: add ESLint dependencies and Convex log reference

* feat: remove obsolete task API and log parameter

* fix: convex client usage for live updates

* fix: use Live Updates for Cleanup Recovery

* fix: remove unused Convex API routes
