--- .github/workflows/keepalive.yml
@@ -1,21 +1,0 @@
-name: Keep Render Alive
-
-on:
-  schedule:
-    - cron: "*/10 * * * *" # Every 10 minutes
-  workflow_dispatch: # Manual trigger option
-
-jobs:
-  ping:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Ping Render Health Endpoint
-        run: |
-          echo "🔄 Pinging Render at $(date)"
-          response=$(curl -s -o /dev/null -w "%{http_code}" https://api.daemondoc.online/health)
-          if [ $response -eq 200 ]; then
-            echo "✓ Server is alive (HTTP $response)"
-          else
-            echo "✗ Server returned HTTP $response"
-            exit 1
-          fi
