+++ convex-server/convex/_generated/ai/ai-files.state.json
@@ -1,6 +1,6 @@
 {
-  "guidelinesHash": "62d72acb9afcc18f658d88dd772f34b5b1da5fa60ef0402e57a784d97c458e57",
+  "guidelinesHash": "533ba2428f2dc572e825555e6e681d2e56e7e757c15a3fdd036a5d705413f020",
   "agentsMdSectionHash": "5934f676ea9a332e7cd4a4f64aa23b59d926e9faca026c758d4b1f87d2101cc3",
   "claudeMdHash": "5934f676ea9a332e7cd4a4f64aa23b59d926e9faca026c758d4b1f87d2101cc3",
-  "agentSkillsSha": "b86618b5c3c4789c9fed98e84bbc34b3e8e70f20"
+  "agentSkillsSha": "6843b65f3cbcee34bb2bc984d444f42ac7ca2a61"
 }
