+++ convex-server/convex/http.ts
@@ -1,6 +1,5 @@
 import { httpRouter } from "convex/server";
 import { httpAction } from "./_generated/server";
-import { api } from "./_generated/api";
 
 const http = httpRouter();
 
@@ -15,18 +14,6 @@ http.route({
         headers: { "Content-Type": "application/json" },
       }
     );
-  }),
-});
-
-http.route({
-  path: "/api/tasks",
-  method: "GET",
-  handler: httpAction(async (ctx, _req) => {
-    const tasks = await ctx.runQuery(api.tasks.get);
-    return new Response(JSON.stringify(tasks), {
-      status: 200,
-      headers: { "Content-Type": "application/json" },
-    });
   }),
 });
