--- convex-server/convex/tasks.ts
@@ -1,17 +1,0 @@
-import { query } from "./_generated/server";
-import { v } from "convex/values";
-
-export const get = query({
-  args: {},
-  handler: async (ctx) => {
-    return await ctx.db.query("tasks").order("asc").take(100);
-  },
-});
-
-export const getById = query({
-  args: { id: v.id("tasks") },
-  handler: async (ctx, args) => {
-    return await ctx.db.get(args.id);
-  },
-});
-
