+++ server/package.json
@@ -7,6 +7,7 @@
     "dev": "nodemon src/index.js",
     "build": " echo \"No build step required\"",
     "start": "node src/index.js",
+    "lint": "eslint .",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
@@ -28,6 +29,9 @@
     "resend": "^6.9.4"
   },
   "devDependencies": {
+    "@eslint/js": "^9.39.1",
+    "eslint": "^9.39.1",
+    "globals": "^16.5.0",
     "nodemon": "^3.1.11"
   }
 }
