+++ server/src/index.js
@@ -4,7 +4,6 @@ import "dotenv/config";
 import authRoutes from "./routes/auth.routes.js";
 import githubRoutes from "./routes/github.routes.js";
 import emailRoutes from "./routes/email.routes.js";
-import convexRoutes from "./routes/convex.routes.js";
 import { connectDB } from "./db/connectDB.js";
 import { recoverInterruptedCleanupLogs } from "./services/logRecovery.service.js";
 
@@ -28,7 +27,6 @@ app.use(express.json());
 app.use("/auth", authRoutes);
 app.use("/api/github", githubRoutes);
 app.use("/api/email", emailRoutes);
-app.use("/api/convex", convexRoutes);
 
 app.get("/", (req, res) => {
   res.send("Hello from the server!");
